Rename components to reuse and deferentiate between API & APIProduct listing
